+++ src/iso/mod.rs
@@ -1,7 +1,8 @@
 // isobemak/src/iso/mod.rs
 pub mod boot_catalog;
 pub mod dir_record;
-mod iso;
+pub mod iso;
 pub mod volume_descriptor;
 
-pub use self::iso::create_iso_from_img;
+// `create_iso_from_img` is no longer re-exported here.
+// It will be accessed directly via `iso::iso::create_iso_from_img`.
